+++ src/extension.ts
@@ -6,14 +6,12 @@ let originalText: string;
 import { escape } from "querystring";
 import * as vscode from "vscode";
 import MarkdownIt from "markdown-it";
+import { minimatch } from "minimatch";
+import path from "path";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-  // Use the console to output diagnostic information (console.log) and errors (console.error)
-  // This line of code will only be executed once when your extension is activated
-  console.log("HTML & Markdown Shortcuts is now active!");
-
   // The command has been defined in the package.json file
   // Now provide the implementation of the command with registerCommand
   // The commandId parameter must match the command field in package.json
@@ -23,6 +21,18 @@ export function activate(context: vscode.ExtensionContext) {
       // declare local constants
       const editor = vscode.window.activeTextEditor,
         selection = editor?.selection;
+
+      // Check if editor exists and language is in allowed list
+      const config = vscode.workspace.getConfiguration(
+        "html-markdown-shortcuts"
+      );
+      const allowedLanguages = config.get<string[]>("changeCaseLanguages", [
+        "markdown",
+        "html",
+      ]);
+      if (!editor || !allowedLanguages.includes(editor.document.languageId)) {
+        return;
+      }
 
       if (selection && !selection.isEmpty) {
         // declare local constants
@@ -134,9 +144,6 @@ export function activate(context: vscode.ExtensionContext) {
 
       const doc = editor.document;
       if (doc.languageId !== "markdown") {
-        vscode.window.showWarningMessage(
-          "Only Markdown files are supported for now!"
-        );
         return;
       }
 
@@ -204,8 +211,21 @@ export function activate(context: vscode.ExtensionContext) {
   // Auto-update TOC on save for Markdown files
   context.subscriptions.push(
     vscode.workspace.onWillSaveTextDocument((event) => {
-      if (event.document.languageId === "markdown") {
-        vscode.commands.executeCommand("html-markdown-shortcuts.generateToc"); // Reuse the existing generateToc logic
+      const config = vscode.workspace.getConfiguration(
+        "html-markdown-shortcuts"
+      );
+      const autoUpdateToc = config.get<boolean>("autoUpdateToc", true);
+      const excludePatterns = config.get<string[]>("tocExcludePatterns", [
+        "CHANGELOG.md",
+      ]);
+      if (
+        autoUpdateToc &&
+        event.document.languageId === "markdown" &&
+        !excludePatterns.some((pattern) =>
+          minimatch(path.basename(event.document.fileName), pattern)
+        )
+      ) {
+        vscode.commands.executeCommand("html-markdown-shortcuts.generateToc");
       }
     })
   );